asset_test_ui/requirements.txt
streamlit
Pillow
# fastapi